feat: support combined repl field

+++ packages/contracts/contracts/UnirepApp.sol
@@ -8,7 +8,7 @@ import { Unirep } from "@unirep/contracts/Unirep.sol";
 contract UnirepApp {
     Unirep public unirep;
 
-    constructor(Unirep _unirep, uint256 _epochLength) {
+    constructor(Unirep _unirep, uint48 _epochLength) {
         // set unirep address
         unirep = _unirep;
 
@@ -26,7 +26,7 @@ contract UnirepApp {
 
     function submitManyAttestations(
         uint256 epochKey,
-        uint256 targetEpoch,
+        uint48 targetEpoch,
         uint[] calldata fieldIndices,
         uint[] calldata vals
     ) public {
@@ -38,7 +38,7 @@ contract UnirepApp {
 
     function submitAttestation(
         uint256 epochKey,
-        uint256 targetEpoch,
+        uint48 targetEpoch,
         uint256 fieldIndex,
         uint256 val
     ) public {
